healthcheck: Move topoWatch back into TopologyWatcher and use TabletRecorder interface. Move cellsAliases cache into TopologyWatcher
