+++ go/vt/discovery/legacy_topology_watcher.go
@@ -38,7 +38,7 @@ import (
 
 // NewLegacyCellTabletsWatcher returns a LegacyTopologyWatcher that monitors all
 // the tablets in a cell, and starts refreshing.
-func NewLegacyCellTabletsWatcher(ctx context.Context, topoServer *topo.Server, tr TabletRecorder, cell string, refreshInterval time.Duration, refreshKnownTablets bool, topoReadConcurrency int) *LegacyTopologyWatcher {
+func NewLegacyCellTabletsWatcher(ctx context.Context, topoServer *topo.Server, tr LegacyTabletRecorder, cell string, refreshInterval time.Duration, refreshKnownTablets bool, topoReadConcurrency int) *LegacyTopologyWatcher {
 	return NewLegacyTopologyWatcher(ctx, topoServer, tr, cell, refreshInterval, refreshKnownTablets, topoReadConcurrency, func(tw *LegacyTopologyWatcher) ([]*topodatapb.TabletAlias, error) {
 		return tw.topoServer.GetTabletsByCell(ctx, tw.cell)
 	})
@@ -46,7 +46,7 @@ func NewLegacyCellTabletsWatcher(ctx context.Context, topoServer *topo.Server, t
 
 // NewLegacyShardReplicationWatcher returns a LegacyTopologyWatcher that
 // monitors the tablets in a cell/keyspace/shard, and starts refreshing.
-func NewLegacyShardReplicationWatcher(ctx context.Context, topoServer *topo.Server, tr TabletRecorder, cell, keyspace, shard string, refreshInterval time.Duration, topoReadConcurrency int) *LegacyTopologyWatcher {
+func NewLegacyShardReplicationWatcher(ctx context.Context, topoServer *topo.Server, tr LegacyTabletRecorder, cell, keyspace, shard string, refreshInterval time.Duration, topoReadConcurrency int) *LegacyTopologyWatcher {
 	return NewLegacyTopologyWatcher(ctx, topoServer, tr, cell, refreshInterval, true /* RefreshKnownTablets */, topoReadConcurrency, func(tw *LegacyTopologyWatcher) ([]*topodatapb.TabletAlias, error) {
 		sri, err := tw.topoServer.GetShardReplication(ctx, tw.cell, keyspace, shard)
 		switch {
@@ -69,11 +69,11 @@ func NewLegacyShardReplicationWatcher(ctx context.Context, topoServer *topo.Serv
 
 // LegacyTopologyWatcher polls tablet from a configurable set of tablets
 // periodically. When tablets are added / removed, it calls
-// the TabletRecorder AddTablet / RemoveTablet interface appropriately.
+// the LegacyTabletRecorder AddTablet / RemoveTablet interface appropriately.
 type LegacyTopologyWatcher struct {
 	// set at construction time
 	topoServer          *topo.Server
-	tr                  TabletRecorder
+	tr                  LegacyTabletRecorder
 	cell                string
 	refreshInterval     time.Duration
 	refreshKnownTablets bool
@@ -100,7 +100,7 @@ type LegacyTopologyWatcher struct {
 
 // NewLegacyTopologyWatcher returns a LegacyTopologyWatcher that monitors all
 // the tablets in a cell, and starts refreshing.
-func NewLegacyTopologyWatcher(ctx context.Context, topoServer *topo.Server, tr TabletRecorder, cell string, refreshInterval time.Duration, refreshKnownTablets bool, topoReadConcurrency int, getTablets func(tw *LegacyTopologyWatcher) ([]*topodatapb.TabletAlias, error)) *LegacyTopologyWatcher {
+func NewLegacyTopologyWatcher(ctx context.Context, topoServer *topo.Server, tr LegacyTabletRecorder, cell string, refreshInterval time.Duration, refreshKnownTablets bool, topoReadConcurrency int, getTablets func(tw *LegacyTopologyWatcher) ([]*topodatapb.TabletAlias, error)) *LegacyTopologyWatcher {
 	tw := &LegacyTopologyWatcher{
 		topoServer:          topoServer,
 		tr:                  tr,
@@ -121,7 +121,7 @@ func NewLegacyTopologyWatcher(ctx context.Context, topoServer *topo.Server, tr T
 	return tw
 }
 
-// watch polls all tablets and notifies TabletRecorder by adding/removing tablets.
+// watch polls all tablets and notifies LegacyTabletRecorder by adding/removing tablets.
 func (tw *LegacyTopologyWatcher) watch() {
 	defer tw.wg.Done()
 	ticker := time.NewTicker(tw.refreshInterval)
@@ -136,7 +136,7 @@ func (tw *LegacyTopologyWatcher) watch() {
 	}
 }
 
-// loadTablets reads all tablets from topology, and updates TabletRecorder.
+// loadTablets reads all tablets from topology, and updates LegacyTabletRecorder.
 func (tw *LegacyTopologyWatcher) loadTablets() {
 	var wg sync.WaitGroup
 	newTablets := make(map[string]*tabletInfo)
@@ -263,7 +263,7 @@ func (tw *LegacyTopologyWatcher) loadTablets() {
 }
 
 // WaitForInitialTopology waits until the watcher reads all of the topology data
-// for the first time and transfers the information to TabletRecorder via its
+// for the first time and transfers the information to LegacyTabletRecorder via its
 // AddTablet() method.
 func (tw *LegacyTopologyWatcher) WaitForInitialTopology() error {
 	select {
@@ -274,7 +274,7 @@ func (tw *LegacyTopologyWatcher) WaitForInitialTopology() error {
 	}
 }
 
-// Stop stops the watcher. It does not clean up the tablets added to TabletRecorder.
+// Stop stops the watcher. It does not clean up the tablets added to LegacyTabletRecorder.
 func (tw *LegacyTopologyWatcher) Stop() {
 	tw.cancelFunc()
 	// wait for watch goroutine to finish.
@@ -297,22 +297,22 @@ func (tw *LegacyTopologyWatcher) TopoChecksum() uint32 {
 	return tw.topoChecksum
 }
 
-// LegacyFilterByShard is a TabletRecorder filter that filters tablets by
+// LegacyFilterByShard is a LegacyTabletRecorder filter that filters tablets by
 // keyspace/shard.
 type LegacyFilterByShard struct {
-	// tr is the underlying TabletRecorder to forward requests too
-	tr TabletRecorder
+	// tr is the underlying LegacyTabletRecorder to forward requests too
+	tr LegacyTabletRecorder
 
 	// filters is a map of keyspace to filters for shards
 	filters map[string][]*filterShard
 }
 
 // NewLegacyFilterByShard creates a new LegacyFilterByShard on top of an existing
-// TabletRecorder. Each filter is a keyspace|shard entry, where shard
+// LegacyTabletRecorder. Each filter is a keyspace|shard entry, where shard
 // can either be a shard name, or a keyrange. All tablets that match
 // at least one keyspace|shard tuple will be forwarded to the
-// underlying TabletRecorder.
-func NewLegacyFilterByShard(tr TabletRecorder, filters []string) (*LegacyFilterByShard, error) {
+// underlying LegacyTabletRecorder.
+func NewLegacyFilterByShard(tr LegacyTabletRecorder, filters []string) (*LegacyFilterByShard, error) {
 	m := make(map[string][]*filterShard)
 	for _, filter := range filters {
 		parts := strings.Split(filter, "|")
@@ -349,21 +349,21 @@ func NewLegacyFilterByShard(tr TabletRecorder, filters []string) (*LegacyFilterB
 	}, nil
 }
 
-// AddTablet is part of the TabletRecorder interface.
+// AddTablet is part of the LegacyTabletRecorder interface.
 func (fbs *LegacyFilterByShard) AddTablet(tablet *topodatapb.Tablet, name string) {
 	if fbs.isIncluded(tablet) {
 		fbs.tr.AddTablet(tablet, name)
 	}
 }
 
-// RemoveTablet is part of the TabletRecorder interface.
+// RemoveTablet is part of the LegacyTabletRecorder interface.
 func (fbs *LegacyFilterByShard) RemoveTablet(tablet *topodatapb.Tablet) {
 	if fbs.isIncluded(tablet) {
 		fbs.tr.RemoveTablet(tablet)
 	}
 }
 
-// ReplaceTablet is part of the TabletRecorder interface.
+// ReplaceTablet is part of the LegacyTabletRecorder interface.
 func (fbs *LegacyFilterByShard) ReplaceTablet(old, new *topodatapb.Tablet, name string) {
 	if fbs.isIncluded(old) && fbs.isIncluded(new) {
 		fbs.tr.ReplaceTablet(old, new, name)
@@ -371,7 +371,7 @@ func (fbs *LegacyFilterByShard) ReplaceTablet(old, new *topodatapb.Tablet, name 
 }
 
 // isIncluded returns true iff the tablet's keyspace and shard should be
-// forwarded to the underlying TabletRecorder.
+// forwarded to the underlying LegacyTabletRecorder.
 func (fbs *LegacyFilterByShard) isIncluded(tablet *topodatapb.Tablet) bool {
 	canonical, kr, err := topo.ValidateShardName(tablet.Shard)
 	if err != nil {
@@ -392,18 +392,18 @@ func (fbs *LegacyFilterByShard) isIncluded(tablet *topodatapb.Tablet) bool {
 	return false
 }
 
-// LegacyFilterByKeyspace is a TabletRecorder filter that filters tablets by
+// LegacyFilterByKeyspace is a LegacyTabletRecorder filter that filters tablets by
 // keyspace
 type LegacyFilterByKeyspace struct {
-	tr TabletRecorder
+	tr LegacyTabletRecorder
 
 	keyspaces map[string]bool
 }
 
 // NewLegacyFilterByKeyspace creates a new LegacyFilterByKeyspace on top of an existing
-// TabletRecorder. Each filter is a keyspace entry. All tablets that match
-// a keyspace will be forwarded to the underlying TabletRecorder.
-func NewLegacyFilterByKeyspace(tr TabletRecorder, selectedKeyspaces []string) *LegacyFilterByKeyspace {
+// LegacyTabletRecorder. Each filter is a keyspace entry. All tablets that match
+// a keyspace will be forwarded to the underlying LegacyTabletRecorder.
+func NewLegacyFilterByKeyspace(tr LegacyTabletRecorder, selectedKeyspaces []string) *LegacyFilterByKeyspace {
 	m := make(map[string]bool)
 	for _, keyspace := range selectedKeyspaces {
 		m[keyspace] = true
@@ -415,21 +415,21 @@ func NewLegacyFilterByKeyspace(tr TabletRecorder, selectedKeyspaces []string) *L
 	}
 }
 
-// AddTablet is part of the TabletRecorder interface.
+// AddTablet is part of the LegacyTabletRecorder interface.
 func (fbk *LegacyFilterByKeyspace) AddTablet(tablet *topodatapb.Tablet, name string) {
 	if fbk.isIncluded(tablet) {
 		fbk.tr.AddTablet(tablet, name)
 	}
 }
 
-// RemoveTablet is part of the TabletRecorder interface.
+// RemoveTablet is part of the LegacyTabletRecorder interface.
 func (fbk *LegacyFilterByKeyspace) RemoveTablet(tablet *topodatapb.Tablet) {
 	if fbk.isIncluded(tablet) {
 		fbk.tr.RemoveTablet(tablet)
 	}
 }
 
-// ReplaceTablet is part of the TabletRecorder interface.
+// ReplaceTablet is part of the LegacyTabletRecorder interface.
 func (fbk *LegacyFilterByKeyspace) ReplaceTablet(old *topodatapb.Tablet, new *topodatapb.Tablet, name string) {
 	if old.Keyspace != new.Keyspace {
 		log.Errorf("Error replacing old tablet in %v with new tablet in %v", old.Keyspace, new.Keyspace)
@@ -442,7 +442,7 @@ func (fbk *LegacyFilterByKeyspace) ReplaceTablet(old *topodatapb.Tablet, new *to
 }
 
 // isIncluded returns true if the tablet's keyspace should be
-// forwarded to the underlying TabletRecorder.
+// forwarded to the underlying LegacyTabletRecorder.
 func (fbk *LegacyFilterByKeyspace) isIncluded(tablet *topodatapb.Tablet) bool {
 	_, exist := fbk.keyspaces[tablet.Keyspace]
 	return exist
