+++ go/vt/discovery/topology_watcher.go
@@ -17,10 +17,15 @@ limitations under the License.
 package discovery
 
 import (
+	"bytes"
 	"fmt"
+	"hash/crc32"
+	"sort"
 	"strings"
 	"sync"
 	"time"
+
+	"vitess.io/vitess/go/vt/topo/topoproto"
 
 	"vitess.io/vitess/go/vt/key"
 
@@ -57,18 +62,19 @@ type tabletInfo struct {
 
 // NewCellTabletsWatcher returns a TopologyWatcher that monitors all
 // the tablets in a cell, and starts refreshing.
-func NewCellTabletsWatcher(ctx context.Context, topoServer *topo.Server, f TabletFilter, cell string, refreshInterval time.Duration, refreshKnownTablets bool, topoReadConcurrency int) *TopologyWatcher {
-	return NewTopologyWatcher(ctx, topoServer, f, cell, refreshInterval, refreshKnownTablets, topoReadConcurrency, func(tw *TopologyWatcher) ([]*topodatapb.TabletAlias, error) {
+func NewCellTabletsWatcher(ctx context.Context, topoServer *topo.Server, tr TabletRecorder, f TabletFilter, cell string, refreshInterval time.Duration, refreshKnownTablets bool, topoReadConcurrency int) *TopologyWatcher {
+	return NewTopologyWatcher(ctx, topoServer, tr, f, cell, refreshInterval, refreshKnownTablets, topoReadConcurrency, func(tw *TopologyWatcher) ([]*topodatapb.TabletAlias, error) {
 		return tw.topoServer.GetTabletsByCell(ctx, tw.cell)
 	})
 }
 
 // TopologyWatcher polls tablet from a configurable set of tablets
 // periodically. When tablets are added / removed, it calls
-// the TabletRecorder AddTablet / RemoveTablet interface appropriately.
+// the LegacyTabletRecorder AddTablet / RemoveTablet interface appropriately.
 type TopologyWatcher struct {
 	// set at construction time
 	topoServer          *topo.Server
+	tabletRecorder      TabletRecorder
 	tabletFilter        TabletFilter
 	cell                string
 	refreshInterval     time.Duration
@@ -92,13 +98,16 @@ type TopologyWatcher struct {
 	firstLoadDone bool
 	// firstLoadChan is closed when the initial loading of topology data is done.
 	firstLoadChan chan struct{}
+	// cellAliases is a cache of cell aliases
+	cellAliases map[string]string
 }
 
 // NewTopologyWatcher returns a TopologyWatcher that monitors all
 // the tablets in a cell, and starts refreshing.
-func NewTopologyWatcher(ctx context.Context, topoServer *topo.Server, filter TabletFilter, cell string, refreshInterval time.Duration, refreshKnownTablets bool, topoReadConcurrency int, getTablets func(tw *TopologyWatcher) ([]*topodatapb.TabletAlias, error)) *TopologyWatcher {
+func NewTopologyWatcher(ctx context.Context, topoServer *topo.Server, tr TabletRecorder, filter TabletFilter, cell string, refreshInterval time.Duration, refreshKnownTablets bool, topoReadConcurrency int, getTablets func(tw *TopologyWatcher) ([]*topodatapb.TabletAlias, error)) *TopologyWatcher {
 	tw := &TopologyWatcher{
 		topoServer:          topoServer,
+		tabletRecorder:      tr,
 		tabletFilter:        filter,
 		cell:                cell,
 		refreshInterval:     refreshInterval,
@@ -106,6 +115,7 @@ func NewTopologyWatcher(ctx context.Context, topoServer *topo.Server, filter Tab
 		getTablets:          getTablets,
 		sem:                 make(chan int, topoReadConcurrency),
 		tablets:             make(map[string]*tabletInfo),
+		cellAliases:         make(map[string]string),
 	}
 	tw.firstLoadChan = make(chan struct{})
 
@@ -116,7 +126,7 @@ func NewTopologyWatcher(ctx context.Context, topoServer *topo.Server, filter Tab
 }
 
 // WaitForInitialTopology waits until the watcher reads all of the topology data
-// for the first time and transfers the information to TabletRecorder via its
+// for the first time and transfers the information to LegacyTabletRecorder via its
 // AddTablet() method.
 func (tw *TopologyWatcher) WaitForInitialTopology() error {
 	select {
@@ -127,7 +137,7 @@ func (tw *TopologyWatcher) WaitForInitialTopology() error {
 	}
 }
 
-// Stop stops the watcher. It does not clean up the tablets added to TabletRecorder.
+// Stop stops the watcher. It does not clean up the tablets added to LegacyTabletRecorder.
 func (tw *TopologyWatcher) Stop() {
 	tw.cancelFunc()
 	// wait for watch goroutine to finish.
@@ -148,6 +158,177 @@ func (tw *TopologyWatcher) TopoChecksum() uint32 {
 	defer tw.mu.Unlock()
 
 	return tw.topoChecksum
+}
+
+func (tw *TopologyWatcher) watchTopo() {
+	tw.wg.Add(1)
+	defer tw.wg.Done()
+	ticker := time.NewTicker(tw.refreshInterval)
+	defer ticker.Stop()
+	for {
+		tw.loadTablets()
+		select {
+		case <-tw.ctx.Done():
+			return
+		case <-ticker.C:
+		}
+	}
+}
+
+func (tw *TopologyWatcher) loadTablets() {
+	var wg sync.WaitGroup
+	newTablets := make(map[string]*tabletInfo)
+	replacedTablets := make(map[string]*tabletInfo)
+
+	tabletAliases, err := tw.getTablets(tw)
+	topologyWatcherOperations.Add(topologyWatcherOpListTablets, 1)
+	if err != nil {
+		topologyWatcherErrors.Add(topologyWatcherOpListTablets, 1)
+		select {
+		case <-tw.ctx.Done():
+			return
+		default:
+		}
+		log.Errorf("cannot get tablets for cell: %v: %v", tw.cell, err)
+		return
+	}
+
+	// Accumulate a list of all known alias strings to use later
+	// when sorting
+	tabletAliasStrs := make([]string, 0, len(tabletAliases))
+
+	tw.mu.Lock()
+	for _, tAlias := range tabletAliases {
+		aliasStr := topoproto.TabletAliasString(tAlias)
+		tabletAliasStrs = append(tabletAliasStrs, aliasStr)
+
+		if !tw.refreshKnownTablets {
+			if val, ok := tw.tablets[aliasStr]; ok {
+				newTablets[aliasStr] = val
+				continue
+			}
+		}
+
+		wg.Add(1)
+		go func(alias *topodatapb.TabletAlias) {
+			defer wg.Done()
+			tw.sem <- 1 // Wait for active queue to drain.
+			tablet, err := tw.topoServer.GetTablet(tw.ctx, alias)
+			topologyWatcherOperations.Add(topologyWatcherOpGetTablet, 1)
+			<-tw.sem // Done; enable next request to run
+			if err != nil {
+				topologyWatcherErrors.Add(topologyWatcherOpGetTablet, 1)
+				select {
+				case <-tw.ctx.Done():
+					return
+				default:
+				}
+				log.Errorf("cannot get tablet for alias %v: %v", alias, err)
+				return
+			}
+			if !(tw.isTabletInCell(tablet.Tablet) && (tw.tabletFilter == nil || tw.tabletFilter.IsIncluded(tablet.Tablet))) {
+				return
+			}
+			tw.mu.Lock()
+			aliasStr := topoproto.TabletAliasString(alias)
+			newTablets[aliasStr] = &tabletInfo{
+				alias:  aliasStr,
+				key:    TabletToMapKey(tablet.Tablet),
+				tablet: tablet.Tablet,
+			}
+			tw.mu.Unlock()
+		}(tAlias)
+	}
+
+	tw.mu.Unlock()
+	wg.Wait()
+	tw.mu.Lock()
+
+	for alias, newVal := range newTablets {
+		if val, ok := tw.tablets[alias]; !ok {
+			// Check if there's a tablet with the same address key but a
+			// different alias. If so, replace it and keep track of the
+			// replaced alias to make sure it isn't removed later.
+			found := false
+			for _, otherVal := range tw.tablets {
+				if newVal.key == otherVal.key {
+					found = true
+					tw.tabletRecorder.ReplaceTablet(otherVal.tablet, newVal.tablet)
+					topologyWatcherOperations.Add(topologyWatcherOpReplaceTablet, 1)
+					replacedTablets[otherVal.alias] = newVal
+				}
+			}
+			if !found {
+				tw.tabletRecorder.AddTablet(newVal.tablet)
+				topologyWatcherOperations.Add(topologyWatcherOpAddTablet, 1)
+			}
+
+		} else if val.key != newVal.key {
+			// Handle the case where the same tablet alias is now reporting
+			// a different address key.
+			replacedTablets[alias] = newVal
+			tw.tabletRecorder.ReplaceTablet(val.tablet, newVal.tablet)
+			topologyWatcherOperations.Add(topologyWatcherOpReplaceTablet, 1)
+		}
+	}
+
+	for _, val := range tw.tablets {
+		if _, ok := newTablets[val.alias]; !ok {
+			if _, ok2 := replacedTablets[val.alias]; !ok2 {
+				tw.tabletRecorder.RemoveTablet(val.tablet)
+				topologyWatcherOperations.Add(topologyWatcherOpRemoveTablet, 1)
+			}
+		}
+	}
+	tw.tablets = newTablets
+	if !tw.firstLoadDone {
+		tw.firstLoadDone = true
+		close(tw.firstLoadChan)
+	}
+
+	// iterate through the tablets in a stable order and compute a
+	// checksum of the tablet map
+	sort.Strings(tabletAliasStrs)
+	var buf bytes.Buffer
+	for _, alias := range tabletAliasStrs {
+		tabletInfo, ok := tw.tablets[alias]
+		if ok {
+			buf.WriteString(alias)
+			buf.WriteString(tabletInfo.key)
+		}
+	}
+	tw.topoChecksum = crc32.ChecksumIEEE(buf.Bytes())
+	tw.lastRefresh = time.Now()
+
+	tw.mu.Unlock()
+
+}
+
+func (tw *TopologyWatcher) getAliasByCell(cell string) string {
+	tw.mu.Lock()
+	defer tw.mu.Unlock()
+
+	if alias, ok := tw.cellAliases[cell]; ok {
+		return alias
+	}
+
+	alias := topo.GetAliasByCell(context.Background(), tw.topoServer, cell)
+	tw.cellAliases[cell] = alias
+
+	return alias
+}
+
+func (tw *TopologyWatcher) isTabletInCell(tablet *topodatapb.Tablet) bool {
+	if tablet.Type == topodatapb.TabletType_MASTER {
+		return true
+	}
+	if tablet.Alias.Cell == tw.cell {
+		return true
+	}
+	if tw.getAliasByCell(tablet.Alias.Cell) == tw.getAliasByCell(tw.cell) {
+		return true
+	}
+	return false
 }
 
 // TabletFilter is an interface that can be given to a TopologyWatcher
@@ -173,10 +354,10 @@ type filterShard struct {
 }
 
 // NewFilterByShard creates a new FilterByShard on top of an existing
-// TabletRecorder. Each filter is a keyspace|shard entry, where shard
+// LegacyTabletRecorder. Each filter is a keyspace|shard entry, where shard
 // can either be a shard name, or a keyrange. All tablets that match
 // at least one keyspace|shard tuple will be forwarded to the
-// underlying TabletRecorder.
+// underlying LegacyTabletRecorder.
 func NewFilterByShard(filters []string) (*FilterByShard, error) {
 	m := make(map[string][]*filterShard)
 	for _, filter := range filters {
@@ -214,7 +395,7 @@ func NewFilterByShard(filters []string) (*FilterByShard, error) {
 }
 
 // IsIncluded returns true iff the tablet's keyspace and shard should be
-// forwarded to the underlying TabletRecorder.
+// forwarded to the underlying LegacyTabletRecorder.
 func (fbs *FilterByShard) IsIncluded(tablet *topodatapb.Tablet) bool {
 	canonical, kr, err := topo.ValidateShardName(tablet.Shard)
 	if err != nil {
@@ -243,7 +424,7 @@ type FilterByKeyspace struct {
 
 // NewFilterByKeyspace creates a new FilterByKeyspace.
 // Each filter is a keyspace entry. All tablets that match
-// a keyspace will be forwarded to the underlying TabletRecorder.
+// a keyspace will be forwarded to the underlying LegacyTabletRecorder.
 func NewFilterByKeyspace(cell string, selectedKeyspaces []string) *FilterByKeyspace {
 	m := make(map[string]bool)
 	for _, keyspace := range selectedKeyspaces {
@@ -256,7 +437,7 @@ func NewFilterByKeyspace(cell string, selectedKeyspaces []string) *FilterByKeysp
 }
 
 // IsIncluded returns true if the tablet's keyspace should be
-// forwarded to the underlying TabletRecorder.
+// forwarded to the underlying LegacyTabletRecorder.
 func (fbk *FilterByKeyspace) IsIncluded(tablet *topodatapb.Tablet) bool {
 	_, exist := fbk.keyspaces[tablet.Keyspace]
 	return exist
